test: tests for creating note were separated into their own file
